Use DesiredBalanceMetrics rather than adding a new class
